+++ plugins/scaffolder-backend/src/lib/templating/filters/projectSlug/index.ts
@@ -13,13 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {
-  createTemplateFilter,
-  parseRepoUrl,
-} from '@backstage/plugin-scaffolder-node';
-import schema from './schema';
-import examples from './examples';
 import { ScmIntegrations } from '@backstage/integration';
+import { parseRepoUrl } from '@backstage/plugin-scaffolder-node';
+import { createTemplateFilter } from '@backstage/plugin-scaffolder-node/alpha';
+import examples from './examples';
+import schema from './schema';
 
 export default (integrations: ScmIntegrations) =>
   createTemplateFilter({
